refactor: change package structure QUZ-53
