+++ game-service/game-domain/src/main/kotlin/com/grepp/quizy/game/domain/RoomAppender.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.domain.game
+package com.grepp.quizy.game.domain
 
 import org.springframework.stereotype.Component
 import java.util.*
