+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/redis/repository/RoomRedisRepository.kt
@@ -1,6 +1,6 @@
-package com.grepp.quizy.infra.game.repository
+package com.grepp.quizy.game.infra.redis.repository
 
-import com.grepp.quizy.infra.game.redis.entity.RoomRedisEntity
+import com.grepp.quizy.game.infra.redis.entity.RoomRedisEntity
 import org.springframework.data.repository.CrudRepository
 
 interface RoomRedisRepository : CrudRepository<RoomRedisEntity, String>
